Optimization - Initial pdf loading improved by profiler   - PdfStream - lazy loading by StreamSegment   - Suppres Annotations changes during load   - Replace SortedDictionary by Dictionary on
- Images
  - Return force  jpeg loading by SkiaSharp (poor performance of ImageLoader on brawser)
  - Fix ImageLoader defailt ImageInfo.ColorType
  - InlineImage lazyLoading/noDublicate data

- Quad little optimization
- Basic steps for Trimming
- Some fonts actualization from PdfBox
- Some refactoring of documntation
- Fix some samples

+++ dotNET/PdfClown/Documents/Contents/Fonts/TTF/Table/GSUB/LookupTypeLigatureSubstitutionSubstFormat1.cs
@@ -35,13 +35,13 @@ namespace PdfClown.Documents.Contents.Fonts.TTF.Table.GSUB
     {
         private readonly LigatureSetTable[] ligatureSetTables;
 
-        public LookupTypeLigatureSubstitutionSubstFormat1(int substFormat, CoverageTable coverageTable, LigatureSetTable[] ligatureSetTables)
+        public LookupTypeLigatureSubstitutionSubstFormat1(ushort substFormat, CoverageTable coverageTable, LigatureSetTable[] ligatureSetTables)
             : base(substFormat, coverageTable)
         {
             this.ligatureSetTables = ligatureSetTables;
         }
 
-        public override int DoSubstitution(int gid, int coverageIndex)
+        public override ushort DoSubstitution(ushort gid, int coverageIndex)
         {
             throw new NotSupportedException("not applicable");
         }
